requirements.txt
torch
sentence-transformers
nltk
psycopg2
python-dotenv
